Deprecate id compression
Adds a deprecation warning for use of `compressed_id?`, `compress_id`
and `uncompress_id`. `ActiveSupport::Deprecation.deprecate_methods`
did not seem to work (presumable because this is a concern?), so I
elected to add a custom deprecation notice to each method instead.

Important to note: `split_id` uses `uncompress_id` internally, which
is unfortunate, since we don't want to flood the logs with such
deprecation notices unnecessarily. We might want to plan carefully how
we back out that behavior.