Support 5.1/5.2 by using bigserial for pre-5.1 migrations
Rails 5.1 changed the default primary key type from integer to
bigint/bigserial in:

rails/rails#26266

With this change, they modified the default compatibility layer for
migration versions less than 5.1 to default to integer if not provided.
Since we want our default to be bigint/bigserial for those old migrations,
we need to prepend our module on the migration class when Rails tries to
find it in the compatilibity stack.

Also suppress messages for migrations in specs.