Use simple testing via ENV variable matrix
Appraisal is nice but it can be simplified.
Additionally, it doesn't currently support the bundler plugin api so
override_gem, plugin, etc., won't work.  We do this in some repositories
so unless there's a reason to stick with appraisal, it's safer
to use a solution that's simpler and more future-proof.

See: https://github.com/thoughtbot/appraisal/issues 178