Use core minimum but allow 7.0/7.1/7.2 since core sets upper limit
Drop end of life rails and add new versions.
Ensure we're testing and supporting community supported
versions.

Drop:

rails 6.0
rails 6.1

Add:

ruby 3.2
ruby 3.3
rails 7.1
rails 7.2

Note: ruby 2.7, and 3.0 are not difficult to keep maintaining, this repo has fast tests and
rails 7 is the minimum and it requires ruby 2.7+ so it's important to target the rubies supported
by the dependency this repo tracks most closely.